Restructured users controller
